+++ inst/include/rma_background4.h
@@ -30,7 +30,7 @@
  *
  */
 
-void rma_bg_parameters(double *PM,double *param, int rows, int cols, int column);
+void rma_bg_parameters(double *PM,double *param, size_t rows, size_t cols, size_t column);
 
 
 /*! \brief Carryout the RMA background correction for a specified column.
@@ -49,7 +49,7 @@ void rma_bg_parameters(double *PM,double *param, int rows, int cols, int column)
  *
  */
 
-void rma_bg_adjust(double *PM, double *param, int rows, int cols, int column);
+void rma_bg_adjust(double *PM, double *param, size_t rows, size_t cols, size_t column);
 
 /*! \brief Carryout the RMA background correction for each column of a matrix
  *
@@ -67,7 +67,7 @@ void rma_bg_adjust(double *PM, double *param, int rows, int cols, int column);
 
 
 
-void rma_bg_correct(double *PM, int rows, int cols);
+void rma_bg_correct(double *PM, size_t rows, size_t cols);
 
 SEXP R_rma_bg_correct(SEXP PMmat,SEXP copy);
